Add Users permission

+++ app/Http/Controllers/PostController.php
@@ -10,7 +10,7 @@ class PostController extends Controller
 {
     public function index()
     {
-        // sleep(3);
+        $this->authorize('viewAny', Post::class);
 
         $posts = PostResource::collection(Post::all());
 
@@ -19,12 +19,14 @@ class PostController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Post::class);
+
         return inertia('Posts/Create');
     }
 
     public function store(StorePostRequest $request)
     {
-        // sleep(3);
+        $this->authorize('create', Post::class);
 
         Post::create($request->validated());
 
@@ -34,11 +36,15 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        $this->authorize('create', Post::class);
+
         return inertia('Posts/Edit', compact('post'));
     }
 
     public function update(Post $post, StorePostRequest $request)
     {
+        $this->authorize('create', Post::class);
+
         $post->update($request->validated());
 
         return redirect()->route('posts.index')
@@ -48,6 +54,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        $this->authorize('create', Post::class);
+        
         $post->delete();
 
         return redirect()->route('posts.index')
